--- a/vllm/block.py
+++ b/vllm/block.py
@@ -7,7 +7,6 @@ _BLANK_TOKEN_ID = -1
 
 DEFAULT_LAST_ACCESSED_TIME = -1
 
-
 class LogicalTokenBlock:
     """A block that stores a contiguous chunk of tokens from left to right.