--- a/vllm/config.py
+++ b/vllm/config.py
@@ -295,12 +295,14 @@ class CacheConfig:
         swap_space: int,
         cache_dtype: str,
         sliding_window: Optional[int] = None,
+        disable_prefix_caching: bool = False,
     ) -> None:
         self.block_size = block_size
         self.gpu_memory_utilization = gpu_memory_utilization
         self.swap_space_bytes = swap_space * _GB
         self.cache_dtype = cache_dtype
         self.sliding_window = sliding_window
+        self.disable_prefix_caching = disable_prefix_caching
         self._verify_args()
         self._verify_cache_dtype()