@jadielam 2024-02-08
Related PR here: https://github.com/vllm-project/vllm/pull/2511
Not exact same work. This one bounds the cache to not run out of memory.