@SageMoore 2024-02-16
I ran some benchmarks to make sure this patch doesn't introduce any regressions. It looks like we have comparable performance, which is good. The script that I used was lifted from the previous prefix caching PR. [https://github.com/vllm-project/vllm/pull/1669#issuecomment-1853210086](https://github.com/vllm-project/vllm/pull/1669#issuecomment-1853210086) 

Here are the results

On one GPU
Before:
```Processed prompts: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:00<00:00, 21.07it/s]
Processed prompts: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 50/50 [00:00<00:00, 113.95it/s]
cost time 0.5295777320861816
```
After:
```Processed prompts: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:00<00:00, 20.82it/s]
Processed prompts: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 50/50 [00:00<00:00, 110.20it/s]
cost time 0.544050931930542
```

On two GPUs
Before:
```warmup
Processed prompts: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:00<00:00,  6.96it/s]
Processed prompts: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 50/50 [00:01<00:00, 49.09it/s]
cost time 1.231238842010498
```

After:
```Processed prompts: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:00<00:00,  6.98it/s]
Processed prompts: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 50/50 [00:01<00:00, 48.74it/s]
cost time 1.2402698993682861
```

There's a decent amount of variance, +- 5-7 it/s, on my computer. It does seem, however, that this patch is maybe 1-3it/s slower on average. I'll do some additional investigation to make sure we aren't recomputing the last prefix block or something like that. These were run with 125m OPT. I'm happy to try other models if folks are interested. 

Link to the model used: [https://huggingface.co/facebook/opt-125m](https://huggingface.co/facebook/opt-125m)