@SageMoore 2024-02-21
> In general, one thing missing is that we need a way to turn off automatic prefix caching.

For turning off automatic prefix caching is it sufficient to just make sure that the prefix_prefill kernel doesn't run or should we also disable all block reuse in the BlockAllocator?